refactor: move test fakes
