+++ embrace-android-core/src/test/java/io/embrace/android/embracesdk/fakes/FakeVersionChecker.kt
