+++ embrace-android-instrumentation-crash-ndk/src/test/kotlin/io/embrace/android/embracesdk/fakes/FakeMainThreadHandler.kt
