--- embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeActivityTracker.kt
@@ -1,42 +1,0 @@
-package io.embrace.android.embracesdk.fakes
-
-import android.app.Activity
-import android.os.Bundle
-import io.embrace.android.embracesdk.internal.session.lifecycle.ActivityTracker
-
-class FakeActivityTracker(
-    override var foregroundActivity: Activity? = null,
-) : ActivityTracker {
-
-    override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
-        TODO("Not yet implemented")
-    }
-
-    override fun onActivityStarted(activity: Activity) {
-        TODO("Not yet implemented")
-    }
-
-    override fun onActivityResumed(activity: Activity) {
-        TODO("Not yet implemented")
-    }
-
-    override fun onActivityPaused(activity: Activity) {
-        TODO("Not yet implemented")
-    }
-
-    override fun onActivityStopped(activity: Activity) {
-        TODO("Not yet implemented")
-    }
-
-    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {
-        TODO("Not yet implemented")
-    }
-
-    override fun onActivityDestroyed(activity: Activity) {
-        TODO("Not yet implemented")
-    }
-
-    override fun close() {
-        TODO("Not yet implemented")
-    }
-}
