+++ embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/injection/FakePayloadSourceModule.kt
@@ -6,17 +6,22 @@ import io.embrace.android.embracesdk.fakes.FakeHostedSdkVersionInfo
 import io.embrace.android.embracesdk.fakes.FakeLogEnvelopeSource
 import io.embrace.android.embracesdk.fakes.FakeLogPayloadSource
 import io.embrace.android.embracesdk.fakes.FakeMetadataService
-import io.embrace.android.embracesdk.fakes.FakePayloadResurrectionService
 import io.embrace.android.embracesdk.fakes.FakeRnBundleIdTracker
-import io.embrace.android.embracesdk.fakes.FakeSessionEnvelopeSource
 import io.embrace.android.embracesdk.fakes.FakeSessionPayloadSource
 import io.embrace.android.embracesdk.internal.capture.metadata.MetadataService
 import io.embrace.android.embracesdk.internal.envelope.log.LogPayloadSource
+import io.embrace.android.embracesdk.internal.envelope.metadata.EnvelopeMetadataSource
 import io.embrace.android.embracesdk.internal.envelope.metadata.HostedSdkVersionInfo
+import io.embrace.android.embracesdk.internal.envelope.resource.EnvelopeResourceSource
 import io.embrace.android.embracesdk.internal.envelope.session.SessionEnvelopeSource
 import io.embrace.android.embracesdk.internal.envelope.session.SessionPayloadSource
 import io.embrace.android.embracesdk.internal.injection.PayloadSourceModule
+import io.embrace.android.embracesdk.internal.instrumentation.crash.ndk.NativeCrashService
+import io.embrace.android.embracesdk.internal.payload.Envelope
+import io.embrace.android.embracesdk.internal.payload.SessionPayload
 import io.embrace.android.embracesdk.internal.resurrection.PayloadResurrectionService
+import io.embrace.android.embracesdk.internal.session.orchestrator.SessionSnapshotType
+import io.embrace.android.embracesdk.internal.utils.Provider
 
 class FakePayloadSourceModule(
     override val metadataService: MetadataService = FakeMetadataService(),
@@ -41,4 +46,34 @@ class FakePayloadSourceModule(
         envelopeResourceSource,
         logPayloadSource
     )
+}
+
+private class FakePayloadResurrectionService : PayloadResurrectionService {
+
+    var resurrectCount: Int = 0
+
+    override fun resurrectOldPayloads(nativeCrashServiceProvider: Provider<NativeCrashService?>) {
+        resurrectCount++
+    }
+}
+
+private class FakeSessionEnvelopeSource(
+    private val metadataSource: EnvelopeMetadataSource,
+    private val resourceSource: EnvelopeResourceSource,
+    private val sessionPayloadSource: SessionPayloadSource,
+) : SessionEnvelopeSource {
+
+    override fun getEnvelope(
+        endType: SessionSnapshotType,
+        startNewSession: Boolean,
+        crashId: String?,
+    ): Envelope<SessionPayload> {
+        return Envelope(
+            resourceSource.getEnvelopeResource(),
+            metadataSource.getEnvelopeMetadata(),
+            "0.1.0",
+            "spans",
+            sessionPayloadSource.getSessionPayload(endType, startNewSession, crashId)
+        )
+    }
 }
